tests_01: fix - refactor code
